Converted the store to typescript, need to refactor AppState
